fix(analytics): correct AB testing average click pos type

+++ algolia/analytics/responses_ab_testing.go
@@ -45,7 +45,7 @@ type ABTestResponse struct {
 // VariantResponse represents an AB test's variant as returned by the Analytics
 // API when retrieved as part of a response.
 type VariantResponse struct {
-	AverageClickPosition   int                 `json:"averageClickPosition"`
+	AverageClickPosition   float64             `json:"averageClickPosition"`
 	ClickCount             int                 `json:"clickCount"`
 	ClickThroughRate       float64             `json:"clickThroughRate"`
 	ConversionCount        int                 `json:"conversionCount"`
